+++ packages/in_app_purchase/in_app_purchase_android/android/src/main/java/io/flutter/plugins/inapppurchase/MethodCallHandlerImpl.java
@@ -145,6 +145,9 @@ class MethodCallHandlerImpl
       case InAppPurchasePlugin.MethodNames.ACKNOWLEDGE_PURCHASE:
         acknowledgePurchase((String) call.argument("purchaseToken"), result);
         break;
+      case InAppPurchasePlugin.MethodNames.IS_FEATURE_SUPPORTED:
+        isFeatureSupported((String) call.argument("feature"), result);
+        break;
       default:
         result.notImplemented();
     }
@@ -378,5 +381,14 @@ class MethodCallHandlerImpl
 
     result.error("UNAVAILABLE", "BillingClient is unset. Try reconnecting.", null);
     return true;
+  }
+
+  private void isFeatureSupported(String feature, MethodChannel.Result result) {
+    if (billingClientError(result)) {
+      return;
+    }
+    assert billingClient != null;
+    BillingResult billingResult = billingClient.isFeatureSupported(feature);
+    result.success(billingResult.getResponseCode() == BillingClient.BillingResponseCode.OK);
   }
 }