+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,6 +1,14 @@
-## 0.1.3
+## 0.1.3+2
 
 * Fixed typos.
+
+## 0.1.3+1
+
+* Add payment proxy.
+
+## 0.1.3
+
+* Added support for isFeatureSupported in the BillingClientWrapper and in InAppPurchaseAndroidPlatformAddition.
 
 ## 0.1.2